1. chuyen booking ra file khac, 2. doi thanh input datepicker,  3. them Toast thong bao, 4. them animation cho form
