+++ package.json
@@ -6,6 +6,7 @@
         "@emotion/react": "^11.11.0",
         "@emotion/styled": "^11.11.0",
         "@fontsource/roboto": "^5.0.2",
+        "@formkit/auto-animate": "^0.7.0",
         "@material-ui/core": "^4.12.4",
         "@mui/icons-material": "^5.11.16",
         "@mui/material": "^5.13.3",
@@ -14,12 +15,15 @@
         "@testing-library/react": "^13.4.0",
         "@testing-library/user-event": "^13.5.0",
         "axios": "^1.4.0",
+        "framer-motion": "^10.12.18",
         "lodash": "^4.17.21",
         "normalize.css": "^8.0.1",
         "react": "^18.2.0",
+        "react-datepicker": "^4.15.0",
         "react-dom": "^18.2.0",
         "react-router-dom": "^6.14.0",
         "react-scripts": "5.0.1",
+        "react-toastify": "^9.1.3",
         "styled-components": "^5.3.11",
         "web-vitals": "^2.1.4"
     },
